Create maps and summaries from FPBK results.

The main purpose of this function is to take the results from FPBK and make readable maps, a fitted variogram plot, and normal-based prediction intervals. The main input for this function is the output from the FPBK.pred function.

FPBKoutput(pred_info, conf_level = c(0.8, 0.9, 0.95),
  get_krigmap = FALSE, get_sampdetails = FALSE,
  get_variogram = FALSE, pointsize = 2)

Arguments

pred_info
is the output from FPBK.pred in this package.

conf_level
is the desired confidence level for the prediction. If conf_level is a vector, then confidence intervals for each element of the vector will be produced.

get_krigmap
is an indicator for whether or not a grid of the kriged responses is returned

get_sampdetails
is an indicator for whether or not a summary of the sampled counts should be output. This summary includes the total number of animals or plants sited, the total area surveyed, the number of sampled sites, the total number of sites, etc.

get_variogram
is an indicator for whether or not a variogram of the residuals should be returned

pointsize
is the size of the points on the spatial graphic.

Value

prediction for the total with prediction intervals

a map of the kriged counts (optional)

a summary of the sample data (optional)

an empirical variogram from gstat with the fitted variogram model with details of the empirical variogram and spatial parameter estimates (optional)