Estimate Covariance Parameters

Used to estimate spatial covariance parameters for a few different spatial models, including the Exponential, Gaussian, and Spherical. Estimated parameters can then be used in predict to predict unobserved values.

estcovparm(response, designmatrix, xcoordsvec, ycoordsvec,
  CorModel = "Exponential", estmethod = "REML", covestimates = c(NA,
  NA, NA), pivec = NA, Vnn = NA)

Arguments

response
a vector of a response variable, possibly with missing values.

designmatrix
is the matrix of covariates used to regress the response on.

xcoordsvec
is a vector of x coordinates

ycoordsvec
is a vector of y coordinates

CorModel
is the covariance structure. By default, covstruct is

estmethod
is either the default "REML" for restricted maximum likelihood to estimate the covariance parameters and regression coefficients or "ML" to estimate the covariance parameters and regression coefficients. Exponential but other options include the Spherical and the Gaussian.

covestimates
is an optional vector of covariance parameter estimates (nugget, partial sill, range). If these are given and estmethod = "None", the the provided vector are treated as the estimators to create the covariance structure.

pivec
is a vector of estimated detection probabilities on each of the sites

Vnn
is the covariance matrix for the estimated detection probabilities

Value

a list with

a vector of estimated covariance parameters

the estimated covariance matrix for all sites

the QR decomposition

a vector of estimated fixed effects

other components used in slmfit

Details

The function is used internally in slmfit.