Obtains estimates for detection parameters through logistic regression.

Estimates logistic regression coefficients from detection data.

get_detection(formula, data, varmethod = "Bootstrap")

Arguments

formula
is an R linear model formula specifying binary detection as the response variable as well as covariates associated with detection.

data
is the data set with the response column of detection and the covariates to be used in the logistic regression, and the spatial coordinates for all of the sites.

varmethod
is the method to obtain the variance matrix for the detection probabilities, either "Delta" or "Bootstrap".

Value

a list with

Fitted detection coefficients

The asymptotic variance-covariance matrix of the fitted regression coefficients

the formula used to fit the logistic regression model

A list of bootstrapped regression coefficients used to obtain a covariance matrix for detection estimates

The method used to get the variance covariance matrix for detection probabilities.

Examples

df <- data.frame(y = rbinom(30, 1, 0.7), x = rnorm(30, 4, 0.4))
get_detection(formula = y ~ x, data = df)

#> $coefs
#> X(Intercept)           Xx 
#>    10.178233    -2.279482 
#> 
#> $covmat
#>              X(Intercept)        Xx
#> X(Intercept)     32.14014 -7.850180
#> Xx               -7.85018  1.928847
#> 
#> $formula
#> y ~ x
#> <environment: 0x7f88d23d1e88>
#> 
#> $boot.coefs
#>                  [,1]          [,2]
#>    [1,]   10.91639327 -2.471639e+00
#>    [2,]    5.33263369 -9.906644e-01
#>    [3,]   18.78728267 -4.395870e+00
#>    [4,]   14.95093077 -3.534514e+00
#>    [5,]    4.77912230 -1.046117e+00
#>    [6,]    8.48237043 -1.827187e+00
#>    [7,]    0.26038515  2.811483e-01
#>    [8,]   10.65045474 -2.450195e+00
#>    [9,]   13.49876563 -3.053477e+00
#>   [10,]   11.38155626 -2.613145e+00
#>   [11,]   11.55498386 -2.691212e+00
#>   [12,]   16.78467674 -3.684315e+00
#>   [13,]    6.37332547 -1.575249e+00
#>   [14,]   13.54746747 -2.877847e+00
#>   [15,]    8.44257768 -1.761170e+00
#>   [16,]    6.82836481 -1.465860e+00
#>   [17,]    5.97144720 -1.330944e+00
#>   [18,]    2.16862106 -4.791129e-01
#>   [19,]    6.97927054 -1.463193e+00
#>   [20,]    7.81902417 -1.524735e+00
#>   [21,]   20.92172455 -4.865409e+00
#>   [22,]   -1.16160859  5.498808e-01
#>   [23,]    5.59935705 -1.125962e+00
#>   [24,]    5.55275743 -1.140652e+00
#>   [25,]   10.04649778 -2.229747e+00
#>   [26,]   14.57726782 -3.237805e+00
#>   [27,]   16.64800495 -4.005300e+00
#>   [28,]    9.85527872 -2.281640e+00
#>   [29,]   21.64209321 -5.116143e+00
#>   [30,]   15.24709068 -3.382151e+00
#>   [31,]    7.71636663 -1.714475e+00
#>   [32,]    9.40951918 -2.099305e+00
#>   [33,]   11.71742015 -2.599131e+00
#>   [34,]   16.83876791 -3.705341e+00
#>   [35,]   13.97989864 -3.093235e+00
#>   [36,]    9.11606837 -2.012342e+00
#>   [37,]   23.68800671 -5.654206e+00
#>   [38,]   10.47309420 -2.166975e+00
#>   [39,]   20.82269301 -5.086206e+00
#>   [40,]   18.50531144 -4.131346e+00
#>   [41,]    9.79438048 -2.247517e+00
#>   [42,]    5.82347187 -1.167446e+00
#>   [43,]   12.79024129 -2.958836e+00
#>   [44,]    9.25785064 -2.178278e+00
#>   [45,]   13.83977649 -3.280865e+00
#>   [46,]    8.11433909 -1.762585e+00
#>   [47,]   27.93000735 -6.804691e+00
#>   [48,]    4.75565608 -9.823964e-01
#>   [49,]   10.72808445 -2.480206e+00
#>   [50,]   10.83772806 -2.536008e+00
#>   [51,]    6.04294908 -1.203423e+00
#>   [52,]    5.19968774 -1.015253e+00
#>   [53,]   16.22302312 -3.765838e+00
#>   [54,]   15.74193680 -3.575893e+00
#>   [55,]    4.13644033 -6.290467e-01
#>   [56,]   25.70430627 -6.227590e+00
#>   [57,]   15.52796153 -3.651767e+00
#>   [58,]   10.57211824 -2.462275e+00
#>   [59,]   11.09540868 -2.479033e+00
#>   [60,]   11.17680437 -2.482368e+00
#>   [61,]    7.18027263 -1.505520e+00
#>   [62,]    5.05807533 -9.923339e-01
#>   [63,]    4.89480412 -1.017965e+00
#>   [64,]   14.19576667 -3.111348e+00
#>   [65,]   15.05318698 -3.485430e+00
#>   [66,]    1.57903791 -2.217245e-01
#>   [67,]   11.98155087 -2.697406e+00
#>   [68,]   15.80083929 -3.871484e+00
#>   [69,]   10.92152922 -2.513065e+00
#>   [70,]    6.77945248 -1.601260e+00
#>   [71,]   19.76865278 -4.454862e+00
#>   [72,]    4.37959003 -8.273790e-01
#>   [73,]   12.94866230 -3.082900e+00
#>   [74,]   14.20388128 -3.341883e+00
#>   [75,]   16.03246559 -3.754108e+00
#>   [76,]    7.14737070 -1.620489e+00
#>   [77,]   33.19146859 -7.763570e+00
#>   [78,]    0.46991465  1.394599e-01
#>   [79,]    2.43276680 -2.045455e-01
#>   [80,]    5.10477784 -1.088979e+00
#>   [81,]    5.39181395 -9.731941e-01
#>   [82,]    7.69203059 -1.839478e+00
#>   [83,]   24.10375035 -5.912363e+00
#>   [84,]    4.04726028 -5.451976e-01
#>   [85,]   12.55059153 -2.984185e+00
#>   [86,]   10.96529665 -2.728437e+00
#>   [87,]    3.37348185 -4.926350e-01
#>   [88,]   11.85191072 -2.764598e+00
#>   [89,]   35.14521927 -8.446795e+00
#>   [90,]   13.64658204 -3.353278e+00
#>   [91,]    6.90839967 -1.510711e+00
#>   [92,]   10.82894084 -2.429456e+00
#>   [93,]   11.78383359 -2.685218e+00
#>   [94,]   12.91334415 -2.922846e+00
#>   [95,]   13.28789622 -3.212031e+00
#>   [96,]    8.34791048 -1.737952e+00
#>   [97,]    8.54313631 -1.895437e+00
#>   [98,]    9.75387130 -2.272999e+00
#>   [99,]    5.23122208 -9.695941e-01
#>  [100,]    4.37854893 -8.849563e-01
#>  [101,]    8.61593330 -1.870822e+00
#>  [102,]   17.34176733 -4.036586e+00
#>  [103,]   10.71769807 -2.370034e+00
#>  [104,]   11.65395602 -2.654434e+00
#>  [105,]   12.05501739 -2.633755e+00
#>  [106,]   28.66700088 -6.456108e+00
#>  [107,]    8.98008101 -1.950726e+00
#>  [108,]   11.47656382 -2.616659e+00
#>  [109,]   10.68829813 -2.356848e+00
#>  [110,]   13.96143101 -3.306987e+00
#>  [111,]   14.56176808 -3.293901e+00
#>  [112,]   10.99681974 -2.570689e+00
#>  [113,]    7.52092984 -1.669795e+00
#>  [114,]    4.00460818 -7.397792e-01
#>  [115,]    7.72971732 -1.687315e+00
#>  [116,]    9.96379078 -2.268630e+00
#>  [117,]   23.56299097 -5.547460e+00
#>  [118,]   13.75182201 -3.253474e+00
#>  [119,]   12.39770252 -2.887658e+00
#>  [120,]   16.82840237 -4.005447e+00
#>  [121,]    8.41698847 -1.904112e+00
#>  [122,]    2.44560172 -2.153723e-01
#>  [123,]   17.84428902 -4.115378e+00
#>  [124,]   12.49183982 -2.844148e+00
#>  [125,]   18.71983489 -4.459896e+00
#>  [126,]    7.75203653 -1.779929e+00
#>  [127,]    8.88103103 -2.037222e+00
#>  [128,]   17.27041545 -4.204010e+00
#>  [129,]    5.19317360 -9.406916e-01
#>  [130,]    9.33202395 -2.128885e+00
#>  [131,]   13.14617930 -2.848920e+00
#>  [132,]   -0.74438797  5.974454e-01
#>  [133,]   42.76928310 -9.902187e+00
#>  [134,]   10.82985704 -2.306896e+00
#>  [135,]   16.98129462 -4.066823e+00
#>  [136,]    5.91972116 -1.376244e+00
#>  [137,]   23.36051863 -5.550284e+00
#>  [138,]   13.91463930 -3.161893e+00
#>  [139,]    9.75991556 -2.297308e+00
#>  [140,]   16.28294905 -3.686122e+00
#>  [141,]   17.82180597 -4.051937e+00
#>  [142,]   10.86292396 -2.523012e+00
#>  [143,]    4.15964665 -7.866423e-01
#>  [144,]   22.75804877 -5.400919e+00
#>  [145,]   14.29650363 -3.524422e+00
#>  [146,]   11.26159093 -2.367301e+00
#>  [147,]   14.46542184 -3.479724e+00
#>  [148,]    3.02723419 -4.621144e-01
#>  [149,]   12.75405093 -2.807949e+00
#>  [150,]   18.49807261 -4.387848e+00
#>  [151,]   16.16460291 -3.525421e+00
#>  [152,]    6.12593901 -9.577141e-01
#>  [153,]   14.94891968 -3.612874e+00
#>  [154,]   27.21544234 -6.799009e+00
#>  [155,]    7.98759810 -1.903823e+00
#>  [156,]    0.26054697  1.076810e-01
#>  [157,]    9.24687367 -2.048369e+00
#>  [158,]   11.17808567 -2.521691e+00
#>  [159,]    9.17141773 -2.098022e+00
#>  [160,]   22.17722071 -5.035288e+00
#>  [161,]    9.56496836 -2.144583e+00
#>  [162,]   20.08411026 -4.741612e+00
#>  [163,]    8.09061917 -1.708838e+00
#>  [164,]   15.31158976 -3.461166e+00
#>  [165,]   24.45290262 -5.631764e+00
#>  [166,]   12.41149708 -2.970001e+00
#>  [167,]   10.11975944 -2.268464e+00
#>  [168,]    6.84723862 -1.426370e+00
#>  [169,]    7.06065371 -1.628578e+00
#>  [170,]    7.97317504 -1.698264e+00
#>  [171,]   13.03032767 -2.917301e+00
#>  [172,]    7.93217121 -1.761030e+00
#>  [173,]    6.66588169 -1.361322e+00
#>  [174,]   12.42916038 -3.083007e+00
#>  [175,]   18.58858301 -4.435974e+00
#>  [176,]    4.86882840 -1.137022e+00
#>  [177,]   17.54007974 -4.104210e+00
#>  [178,]   16.23184536 -3.887896e+00
#>  [179,]   23.24894109 -5.572330e+00
#>  [180,]    8.47842188 -1.796125e+00
#>  [181,]   11.47557013 -2.667879e+00
#>  [182,]    9.16823950 -2.093079e+00
#>  [183,]   -1.88017169  5.721939e-01
#>  [184,]   20.30364741 -4.744861e+00
#>  [185,]   10.43577705 -2.207606e+00
#>  [186,]   21.80342125 -4.925212e+00
#>  [187,]    6.87879520 -1.363018e+00
#>  [188,]   16.29241897 -3.770868e+00
#>  [189,]    4.22435340 -7.107551e-01
#>  [190,]    5.50145601 -1.068954e+00
#>  [191,]   14.29211593 -3.149250e+00
#>  [192,]    9.06694979 -2.070848e+00
#>  [193,]   18.22535570 -4.220639e+00
#>  [194,]    7.39473826 -1.681124e+00
#>  [195,]    8.14009435 -1.740120e+00
#>  [196,]   11.45110225 -2.582432e+00
#>  [197,]   11.80812450 -2.679433e+00
#>  [198,]   15.86659127 -3.729172e+00
#>  [199,]   16.08786347 -3.915773e+00
#>  [200,]   17.87123919 -3.973425e+00
#>  [201,]   11.29278340 -2.628585e+00
#>  [202,]   10.90212960 -2.401959e+00
#>  [203,]   13.29250051 -3.115999e+00
#>  [204,]   10.62697380 -2.345235e+00
#>  [205,]   -7.05248340  2.230696e+00
#>  [206,]   12.33793911 -2.868207e+00
#>  [207,]   14.19104491 -3.039440e+00
#>  [208,]   12.63307172 -2.683150e+00
#>  [209,]   14.28406412 -3.453207e+00
#>  [210,]   15.27843162 -3.704701e+00
#>  [211,]   18.52883653 -4.278059e+00
#>  [212,]    6.40506207 -1.240127e+00
#>  [213,]   10.51118843 -2.373520e+00
#>  [214,]   14.32153265 -3.317302e+00
#>  [215,]   11.56348705 -2.516384e+00
#>  [216,]   14.06382898 -3.156158e+00
#>  [217,]   -1.68003101  8.954395e-01
#>  [218,]   12.19296182 -2.803977e+00
#>  [219,]    2.39491450 -3.788829e-01
#>  [220,]   14.64529407 -3.343675e+00
#>  [221,]    8.57975663 -1.902319e+00
#>  [222,]    8.12062439 -1.722792e+00
#>  [223,]   20.07847150 -4.643459e+00
#>  [224,]   19.76277440 -4.595620e+00
#>  [225,]    4.38044222 -8.351747e-01
#>  [226,]   17.74179854 -4.201531e+00
#>  [227,]    4.93894165 -8.504451e-01
#>  [228,]   11.80949507 -2.680619e+00
#>  [229,]   14.30290612 -3.299731e+00
#>  [230,]   13.43293329 -3.122798e+00
#>  [231,]   12.59198203 -3.026161e+00
#>  [232,]   15.27785069 -3.760047e+00
#>  [233,]   15.98535214 -3.797683e+00
#>  [234,]    8.09544949 -1.741404e+00
#>  [235,]    6.99878348 -1.606355e+00
#>  [236,]    5.27508425 -1.028791e+00
#>  [237,]    8.71795461 -1.815060e+00
#>  [238,]   10.09982648 -2.358304e+00
#>  [239,]   10.91290089 -2.423998e+00
#>  [240,]   25.26693564 -5.945296e+00
#>  [241,]    3.65933418 -5.779751e-01
#>  [242,]   20.17431454 -4.609657e+00
#>  [243,]   23.70692508 -5.687908e+00
#>  [244,]   17.25691401 -4.136350e+00
#>  [245,]   11.52254021 -2.789135e+00
#>  [246,]   11.39252594 -2.592995e+00
#>  [247,]   12.82458428 -2.937386e+00
#>  [248,]    8.52104062 -1.742386e+00
#>  [249,]   13.89625199 -3.298117e+00
#>  [250,]    5.12018437 -9.932744e-01
#>  [251,]   18.32509767 -4.171451e+00
#>  [252,]   14.26347696 -3.244851e+00
#>  [253,]   11.02569074 -2.366511e+00
#>  [254,]   11.53506414 -2.599103e+00
#>  [255,]   14.64647997 -3.501118e+00
#>  [256,]    7.52544173 -1.656739e+00
#>  [257,]   10.97702263 -2.379943e+00
#>  [258,]    9.56519152 -2.070378e+00
#>  [259,]   10.25662418 -2.284501e+00
#>  [260,]   12.12134778 -2.816169e+00
#>  [261,]    7.20466484 -1.598056e+00
#>  [262,]    8.60495569 -1.836617e+00
#>  [263,]    9.72691408 -2.040339e+00
#>  [264,]   20.13376019 -4.797078e+00
#>  [265,]    7.88013678 -1.810227e+00
#>  [266,]   21.10145336 -5.159240e+00
#>  [267,]    7.75583784 -1.737148e+00
#>  [268,]   17.32573850 -3.926933e+00
#>  [269,]    6.96542700 -1.366316e+00
#>  [270,]   12.25321026 -2.857327e+00
#>  [271,]    8.65645414 -1.771125e+00
#>  [272,]   18.06705563 -4.103635e+00
#>  [273,]   11.70660212 -2.625000e+00
#>  [274,]   16.27292617 -3.964987e+00
#>  [275,]    8.56610833 -2.002320e+00
#>  [276,]   10.28617861 -2.094512e+00
#>  [277,]   18.42247396 -4.260274e+00
#>  [278,]    8.57558036 -1.898066e+00
#>  [279,] 2191.36720582 -5.149939e+02
#>  [280,]   18.49946061 -4.309000e+00
#>  [281,]    3.81267968 -5.491702e-01
#>  [282,]    5.70983392 -1.080056e+00
#>  [283,]    7.86457973 -1.712032e+00
#>  [284,]    9.18801579 -1.943882e+00
#>  [285,]    4.16763366 -8.717932e-01
#>  [286,]    5.07338197 -1.069887e+00
#>  [287,]   17.30400944 -4.048186e+00
#>  [288,]   10.56152061 -2.421377e+00
#>  [289,]   18.43748481 -4.367156e+00
#>  [290,]    5.69374971 -1.113393e+00
#>  [291,]   13.00457253 -2.857297e+00
#>  [292,]   13.08275545 -2.938853e+00
#>  [293,]    7.57829348 -1.765879e+00
#>  [294,]    6.67874093 -1.508278e+00
#>  [295,]   25.65169246 -5.946806e+00
#>  [296,]   14.65526518 -3.330811e+00
#>  [297,]   10.89421582 -2.347561e+00
#>  [298,]    5.34082601 -1.225121e+00
#>  [299,]    9.79141920 -2.363381e+00
#>  [300,]    5.72308823 -1.144925e+00
#>  [301,]   12.90744356 -2.893990e+00
#>  [302,]   39.58879494 -9.633172e+00
#>  [303,]    8.13032520 -1.801646e+00
#>  [304,]   13.20352878 -2.940273e+00
#>  [305,]    5.61637678 -1.181000e+00
#>  [306,]    7.32433977 -1.788707e+00
#>  [307,]   20.34671555 -4.655954e+00
#>  [308,]   16.70918348 -3.534943e+00
#>  [309,]    9.59030792 -2.185864e+00
#>  [310,]   16.37091343 -3.961273e+00
#>  [311,]    8.58090871 -2.155411e+00
#>  [312,]   12.28898859 -2.762239e+00
#>  [313,]    8.61510009 -1.900667e+00
#>  [314,]   20.40065476 -4.721738e+00
#>  [315,]    9.04255322 -1.986592e+00
#>  [316,]    3.00571474 -6.096578e-01
#>  [317,]   13.41156097 -2.911222e+00
#>  [318,]   15.11888986 -3.547983e+00
#>  [319,]   11.39716773 -2.731997e+00
#>  [320,]   16.44603869 -3.677112e+00
#>  [321,]    8.85037690 -2.136363e+00
#>  [322,]    6.46360857 -1.225035e+00
#>  [323,]   12.63854261 -3.051571e+00
#>  [324,]    9.71749716 -2.159979e+00
#>  [325,]    5.69214673 -1.080018e+00
#>  [326,]    7.58864534 -1.563773e+00
#>  [327,]   14.79727289 -3.230889e+00
#>  [328,]   -0.24356283  4.648818e-01
#>  [329,]   12.28210744 -2.742842e+00
#>  [330,]    6.82973281 -1.406699e+00
#>  [331,]    5.56268947 -1.156776e+00
#>  [332,]   14.22978025 -3.344666e+00
#>  [333,]   15.07925068 -3.431554e+00
#>  [334,]   11.39246816 -2.598441e+00
#>  [335,]   12.33199357 -2.818695e+00
#>  [336,]   12.88978095 -2.977345e+00
#>  [337,]    6.63357522 -1.446007e+00
#>  [338,]    7.09174251 -1.492583e+00
#>  [339,]    6.90665112 -1.475051e+00
#>  [340,]    9.12495468 -2.043522e+00
#>  [341,]    0.20636422  4.313724e-01
#>  [342,]    8.64894194 -1.999663e+00
#>  [343,]    7.58385169 -1.721059e+00
#>  [344,]    5.02396525 -9.376277e-01
#>  [345,]    9.50581037 -1.915262e+00
#>  [346,]    8.39220212 -1.903184e+00
#>  [347,]   10.84652246 -2.309650e+00
#>  [348,]   10.41434597 -2.171749e+00
#>  [349,]   21.61758320 -5.299852e+00
#>  [350,]   28.63506628 -6.810777e+00
#>  [351,]   16.98669926 -4.059195e+00
#>  [352,]   20.38468867 -4.829967e+00
#>  [353,]    9.46084488 -1.846209e+00
#>  [354,]    3.53598108 -4.123677e-01
#>  [355,]   14.84954328 -3.412670e+00
#>  [356,]   16.42318390 -3.838314e+00
#>  [357,]   18.90725108 -4.353607e+00
#>  [358,]   12.80674649 -2.866226e+00
#>  [359,]   20.24333732 -4.666545e+00
#>  [360,]    6.19129177 -1.393568e+00
#>  [361,]   16.75179819 -3.987748e+00
#>  [362,]    4.07243953 -7.337196e-01
#>  [363,]   10.53684274 -2.644089e+00
#>  [364,]    6.84439554 -1.369735e+00
#>  [365,]    9.15936951 -1.926278e+00
#>  [366,]   18.55877875 -4.311990e+00
#>  [367,]    7.53404402 -1.607410e+00
#>  [368,]    3.54528293 -6.352720e-01
#>  [369,]   16.13409174 -3.691797e+00
#>  [370,]    9.53031249 -2.037250e+00
#>  [371,]   10.06678757 -2.179117e+00
#>  [372,]    6.94086849 -1.595787e+00
#>  [373,]   11.35027518 -2.573629e+00
#>  [374,]    8.52816887 -1.943260e+00
#>  [375,]   14.80432005 -3.423478e+00
#>  [376,]   11.39056117 -2.666574e+00
#>  [377,]    8.06428999 -1.801934e+00
#>  [378,]   11.31649991 -2.408545e+00
#>  [379,]   19.29911199 -4.339125e+00
#>  [380,]   11.36632978 -2.619168e+00
#>  [381,]    5.03419527 -8.029325e-01
#>  [382,]   19.74613864 -4.653968e+00
#>  [383,]    6.55450791 -1.405209e+00
#>  [384,]    5.44835822 -1.152975e+00
#>  [385,]   10.69443004 -2.376236e+00
#>  [386,]    5.35793551 -1.248048e+00
#>  [387,]    8.88864259 -2.166202e+00
#>  [388,]   10.42671074 -2.226049e+00
#>  [389,]    8.35823289 -1.808139e+00
#>  [390,]    9.73305420 -2.173457e+00
#>  [391,]   20.46711554 -5.236910e+00
#>  [392,]   21.59848898 -5.062224e+00
#>  [393,]    8.96972593 -2.104223e+00
#>  [394,]   11.53585801 -2.565943e+00
#>  [395,]   11.30574838 -2.547765e+00
#>  [396,]   -1.36025001  4.031494e-01
#>  [397,]   19.49043999 -4.568369e+00
#>  [398,]   11.11654558 -2.524842e+00
#>  [399,]   14.63339015 -3.389823e+00
#>  [400,]    6.83875737 -1.152397e+00
#>  [401,]   12.68089148 -3.087504e+00
#>  [402,]    7.08038729 -1.388282e+00
#>  [403,]   -0.30575598  2.873980e-01
#>  [404,]   13.33977046 -3.014120e+00
#>  [405,]   20.62173876 -4.807666e+00
#>  [406,]   17.29481271 -4.140280e+00
#>  [407,]    8.05514223 -1.798336e+00
#>  [408,]   22.68027170 -5.488422e+00
#>  [409,]    9.08671029 -2.065451e+00
#>  [410,]    5.33610220 -1.241662e+00
#>  [411,]    7.50091456 -1.565008e+00
#>  [412,]   10.31262860 -2.370953e+00
#>  [413,]    9.36355905 -2.186934e+00
#>  [414,]    0.84445773  7.095236e-04
#>  [415,]    5.29363253 -1.091832e+00
#>  [416,]   15.74263993 -3.833936e+00
#>  [417,]   13.50791475 -3.317820e+00
#>  [418,]   17.02032054 -3.752511e+00
#>  [419,]   11.50168123 -2.683089e+00
#>  [420,]   -0.60310805  5.589821e-01
#>  [421,]    5.03454768 -1.040386e+00
#>  [422,]    6.73731231 -1.272479e+00
#>  [423,]   22.13508514 -4.909553e+00
#>  [424,]    8.02783995 -1.617852e+00
#>  [425,]    9.94310535 -2.199477e+00
#>  [426,]   11.54657510 -2.565497e+00
#>  [427,]   15.51045533 -3.558936e+00
#>  [428,]    0.92649902  6.689595e-02
#>  [429,]    2.68954692 -4.655709e-01
#>  [430,]   13.84422342 -3.017058e+00
#>  [431,]   19.69316796 -4.564840e+00
#>  [432,]   31.37861455 -7.150841e+00
#>  [433,]   11.20404071 -2.550288e+00
#>  [434,]    7.02151734 -1.506496e+00
#>  [435,]   13.19660803 -2.978020e+00
#>  [436,]   12.44744214 -2.788849e+00
#>  [437,]   17.98061443 -4.196036e+00
#>  [438,]   14.13999301 -3.461030e+00
#>  [439,]    3.99011949 -7.804985e-01
#>  [440,]   10.31014547 -2.272386e+00
#>  [441,]   33.86947276 -7.797843e+00
#>  [442,]    5.92725750 -1.192383e+00
#>  [443,]    3.81401682 -7.912326e-01
#>  [444,]   32.76370228 -8.013519e+00
#>  [445,]   10.00733782 -2.168563e+00
#>  [446,]   21.98060882 -5.280954e+00
#>  [447,]   12.24521122 -2.817703e+00
#>  [448,]    8.69310652 -2.027606e+00
#>  [449,]    2.20267113 -1.525709e-01
#>  [450,]    8.12055201 -1.870837e+00
#>  [451,]   16.72270728 -3.778218e+00
#>  [452,]   11.24476047 -2.368314e+00
#>  [453,]   19.50789825 -4.555615e+00
#>  [454,]   12.59378864 -2.840254e+00
#>  [455,]    9.40045076 -2.083280e+00
#>  [456,]    4.26719249 -4.099462e-01
#>  [457,]   11.45004012 -2.445013e+00
#>  [458,]    1.50388689  1.746731e-01
#>  [459,]    1.26419999  8.678047e-02
#>  [460,]   10.59133058 -2.379504e+00
#>  [461,]   34.64376583 -8.252374e+00
#>  [462,]   23.79888324 -5.658676e+00
#>  [463,]   14.27326482 -3.149502e+00
#>  [464,]   16.43749600 -3.670058e+00
#>  [465,]   20.19851840 -4.452515e+00
#>  [466,]    9.86877248 -2.303023e+00
#>  [467,]   24.25320179 -5.622155e+00
#>  [468,]    6.15253429 -1.385413e+00
#>  [469,]    6.52251106 -1.368121e+00
#>  [470,]   11.00721068 -2.476872e+00
#>  [471,]    8.80152372 -2.000159e+00
#>  [472,]   16.30709203 -3.820368e+00
#>  [473,]   18.22958262 -4.379346e+00
#>  [474,]    9.89481662 -2.017480e+00
#>  [475,]    8.50476371 -1.995585e+00
#>  [476,]   22.71538320 -5.586829e+00
#>  [477,]   13.71905256 -3.137506e+00
#>  [478,]   21.82451343 -5.241107e+00
#>  [479,]    4.36202281 -7.639857e-01
#>  [480,]   13.14525236 -3.026246e+00
#>  [481,]    7.07788012 -1.378935e+00
#>  [482,]    5.25131822 -9.610268e-01
#>  [483,]   16.91121384 -3.690116e+00
#>  [484,]   19.38890195 -4.678168e+00
#>  [485,]    5.98100692 -1.096598e+00
#>  [486,]   17.61754242 -4.026062e+00
#>  [487,]    2.35237695 -2.395334e-01
#>  [488,]    6.36583787 -1.300958e+00
#>  [489,]   20.28238717 -4.957597e+00
#>  [490,]   11.81038119 -2.785529e+00
#>  [491,]   15.53596944 -3.408034e+00
#>  [492,]    8.29078550 -1.779026e+00
#>  [493,]    6.02586839 -1.163934e+00
#>  [494,]   14.50135176 -3.542178e+00
#>  [495,]   12.17115362 -2.659385e+00
#>  [496,]   12.35666098 -2.689759e+00
#>  [497,]    9.36084261 -2.190429e+00
#>  [498,]    8.05535914 -1.671814e+00
#>  [499,]    4.56839885 -7.429320e-01
#>  [500,]   12.28731299 -3.019130e+00
#>  [501,]    5.67260196 -1.085474e+00
#>  [502,]   23.53802363 -5.330857e+00
#>  [503,]    4.29945470 -8.631435e-01
#>  [504,]   11.26770852 -2.436081e+00
#>  [505,]    6.16264921 -1.141042e+00
#>  [506,]   11.15497522 -2.535487e+00
#>  [507,]    9.53111801 -1.979543e+00
#>  [508,]   15.37313918 -3.479115e+00
#>  [509,]   12.86029357 -2.881941e+00
#>  [510,]    8.83239781 -1.955281e+00
#>  [511,]   13.68240461 -3.071591e+00
#>  [512,]    6.62236232 -1.437398e+00
#>  [513,]   13.66417716 -3.162217e+00
#>  [514,]   15.18934157 -3.355681e+00
#>  [515,]   40.09247628 -9.603473e+00
#>  [516,]    7.47710448 -1.741591e+00
#>  [517,]    7.04599092 -1.511059e+00
#>  [518,]    7.74799474 -1.850703e+00
#>  [519,]   14.05815859 -3.355461e+00
#>  [520,]   20.24691697 -4.840708e+00
#>  [521,]   12.32014522 -2.920572e+00
#>  [522,]   24.04078779 -5.504852e+00
#>  [523,]   11.30179437 -2.552330e+00
#>  [524,]    2.35231008 -4.193335e-01
#>  [525,]   29.98037272 -7.531026e+00
#>  [526,]   20.70208510 -5.009907e+00
#>  [527,]    7.43966320 -1.643162e+00
#>  [528,]    6.19996796 -1.224210e+00
#>  [529,]    4.49521803 -8.339761e-01
#>  [530,]   19.22827734 -4.323475e+00
#>  [531,]   17.13500492 -3.945186e+00
#>  [532,]   12.20479628 -2.734255e+00
#>  [533,]   14.10094023 -3.216564e+00
#>  [534,]   15.32053959 -3.677775e+00
#>  [535,]    8.81309933 -1.883616e+00
#>  [536,]   10.65539870 -2.500778e+00
#>  [537,]   19.18699944 -4.603004e+00
#>  [538,]   16.34747586 -3.736849e+00
#>  [539,]    8.10390649 -1.754213e+00
#>  [540,]    8.51755041 -1.727739e+00
#>  [541,]   10.06943279 -2.359871e+00
#>  [542,]   14.07263589 -3.357492e+00
#>  [543,]   14.42556659 -3.264902e+00
#>  [544,]   12.06485916 -2.797846e+00
#>  [545,]   12.13664893 -2.732430e+00
#>  [546,]    8.46623102 -1.857719e+00
#>  [547,]   12.36861689 -2.789312e+00
#>  [548,]   -8.04011926  2.302212e+00
#>  [549,]    7.23861432 -1.348912e+00
#>  [550,]    8.70170009 -1.824394e+00
#>  [551,]   24.74305143 -5.900514e+00
#>  [552,]   10.09487116 -2.179639e+00
#>  [553,]   31.68754248 -7.139501e+00
#>  [554,]   13.23017303 -3.014067e+00
#>  [555,]    6.38130027 -1.541791e+00
#>  [556,]   26.08556209 -6.055721e+00
#>  [557,]   10.43665251 -2.548259e+00
#>  [558,]   18.84901971 -4.544160e+00
#>  [559,]    3.51789583 -5.712851e-01
#>  [560,]    9.42915879 -2.190279e+00
#>  [561,]   22.30039123 -5.199851e+00
#>  [562,]    8.80520046 -1.994796e+00
#>  [563,]   12.19449358 -2.795161e+00
#>  [564,]   28.50228796 -6.487022e+00
#>  [565,]    3.20080170 -5.784939e-01
#>  [566,]   15.80740669 -3.806842e+00
#>  [567,]   14.76629541 -3.525195e+00
#>  [568,]    0.97588715 -7.116355e-02
#>  [569,]   18.35833403 -4.301870e+00
#>  [570,]    6.07180486 -1.333169e+00
#>  [571,]   18.71247586 -4.168771e+00
#>  [572,]   10.83281574 -2.506643e+00
#>  [573,]    8.13339282 -1.822043e+00
#>  [574,]   14.39914525 -3.389340e+00
#>  [575,]   12.51813924 -2.619246e+00
#>  [576,]   17.18097166 -4.139062e+00
#>  [577,]   22.24205841 -5.181381e+00
#>  [578,]    8.13391420 -1.794876e+00
#>  [579,]   10.21295343 -2.263514e+00
#>  [580,]    6.62732441 -1.500012e+00
#>  [581,]    6.30203328 -1.201746e+00
#>  [582,]   16.12256974 -3.828503e+00
#>  [583,]   10.54544699 -2.232996e+00
#>  [584,]   35.49061932 -8.626439e+00
#>  [585,]   11.66879382 -2.361686e+00
#>  [586,]   31.26235513 -7.327026e+00
#>  [587,]   10.34052083 -2.349878e+00
#>  [588,]   10.59955603 -2.344630e+00
#>  [589,]   11.72478154 -2.694149e+00
#>  [590,]   21.44003855 -4.866164e+00
#>  [591,]   12.94459176 -2.980030e+00
#>  [592,]   10.79390553 -2.483221e+00
#>  [593,]   -0.45117879  4.575940e-01
#>  [594,]    8.51654454 -2.043807e+00
#>  [595,]   12.55231537 -2.895543e+00
#>  [596,]    8.53185513 -1.915919e+00
#>  [597,]   11.61327412 -2.764500e+00
#>  [598,]   12.49643837 -2.954989e+00
#>  [599,]   19.22864958 -4.519896e+00
#>  [600,]   11.06949904 -2.444062e+00
#>  [601,]   11.67458506 -2.418798e+00
#>  [602,]    2.92724505 -5.277030e-01
#>  [603,]   11.73912703 -2.794361e+00
#>  [604,]    6.49272428 -1.428802e+00
#>  [605,]   15.65328905 -3.586201e+00
#>  [606,]    9.83602709 -2.175231e+00
#>  [607,]   12.41443660 -2.707401e+00
#>  [608,]   10.82676684 -2.273679e+00
#>  [609,]   11.99010409 -2.906340e+00
#>  [610,]    6.03235822 -1.297763e+00
#>  [611,]   15.35824253 -3.592698e+00
#>  [612,]   13.82293803 -3.399382e+00
#>  [613,]   17.82785357 -4.023605e+00
#>  [614,]    8.09352825 -1.924059e+00
#>  [615,]    2.01071185 -2.548454e-01
#>  [616,]   11.66234945 -2.667900e+00
#>  [617,]    8.44847061 -1.887468e+00
#>  [618,]   10.79438592 -2.423685e+00
#>  [619,]    5.37282935 -1.229716e+00
#>  [620,]    8.67605098 -1.819911e+00
#>  [621,]   19.19876498 -4.338580e+00
#>  [622,]    7.55148121 -1.609671e+00
#>  [623,]   15.15803491 -3.596538e+00
#>  [624,]   13.59332898 -3.180323e+00
#>  [625,]   12.95219098 -2.635783e+00
#>  [626,]   17.38723157 -4.036273e+00
#>  [627,]   25.70437392 -6.261541e+00
#>  [628,]    8.79284538 -1.706927e+00
#>  [629,]    8.88061992 -1.746574e+00
#>  [630,]    8.08013438 -1.787655e+00
#>  [631,]    5.41221064 -1.133967e+00
#>  [632,]    8.67609682 -1.935366e+00
#>  [633,]   10.26796204 -2.294932e+00
#>  [634,]    2.20314930 -8.430021e-02
#>  [635,]   21.90581400 -5.110438e+00
#>  [636,]    3.55653662 -4.990863e-01
#>  [637,]   15.09468385 -3.618123e+00
#>  [638,]   10.41884091 -2.361603e+00
#>  [639,]    6.85757227 -1.377896e+00
#>  [640,]   17.26170730 -3.882384e+00
#>  [641,]   10.98842006 -2.467226e+00
#>  [642,]   12.77466896 -2.901449e+00
#>  [643,]   23.67190682 -5.420276e+00
#>  [644,]   17.55864916 -4.118466e+00
#>  [645,]   15.04090785 -3.587938e+00
#>  [646,]    8.00363369 -1.817911e+00
#>  [647,]    6.10553671 -1.164648e+00
#>  [648,]    3.71069705 -7.236083e-01
#>  [649,]   12.74482743 -2.934560e+00
#>  [650,]   10.59578432 -2.453249e+00
#>  [651,]   17.56962364 -4.164995e+00
#>  [652,]   -1.15206914  4.581048e-01
#>  [653,]   10.27667451 -2.430175e+00
#>  [654,]    9.42844031 -2.044433e+00
#>  [655,]   18.06720116 -4.138362e+00
#>  [656,]    7.49545353 -1.727111e+00
#>  [657,]    5.00416231 -8.760695e-01
#>  [658,]   10.17895961 -2.426518e+00
#>  [659,]   17.27307147 -4.050090e+00
#>  [660,]   21.74427427 -4.774252e+00
#>  [661,]   17.90541901 -4.284671e+00
#>  [662,]   16.38329539 -3.801788e+00
#>  [663,]    4.92414762 -6.912417e-01
#>  [664,]    8.75745994 -1.924749e+00
#>  [665,]   29.13199657 -7.036228e+00
#>  [666,]   17.48999436 -4.186954e+00
#>  [667,]   11.79367781 -2.741734e+00
#>  [668,]   17.62079850 -4.215103e+00
#>  [669,]   -0.38423063  3.989916e-01
#>  [670,]    4.70592658 -8.479189e-01
#>  [671,]   15.98544149 -3.969272e+00
#>  [672,]   12.53865210 -2.753524e+00
#>  [673,]    6.22153386 -1.121222e+00
#>  [674,]   24.74300766 -6.112293e+00
#>  [675,]   10.59594894 -2.404740e+00
#>  [676,]   18.14394704 -4.054984e+00
#>  [677,]    8.34568222 -1.613353e+00
#>  [678,]   15.09616657 -3.571175e+00
#>  [679,]    7.36910791 -1.816786e+00
#>  [680,]   14.69495924 -3.549374e+00
#>  [681,]   11.66803178 -2.719913e+00
#>  [682,]   14.54995763 -3.199691e+00
#>  [683,]   17.47918903 -4.051405e+00
#>  [684,]   18.85587760 -4.206202e+00
#>  [685,]    8.17881631 -1.807400e+00
#>  [686,]   12.17860783 -2.917403e+00
#>  [687,]   12.13279101 -2.749217e+00
#>  [688,]    8.63989662 -1.936011e+00
#>  [689,]   10.59491160 -2.320414e+00
#>  [690,]   13.59923440 -3.215851e+00
#>  [691,]   12.45411702 -2.891999e+00
#>  [692,]   24.38165709 -5.428211e+00
#>  [693,]   17.97657013 -3.962942e+00
#>  [694,]   13.82462805 -3.341490e+00
#>  [695,]    5.75658354 -1.212496e+00
#>  [696,]    7.24103706 -1.506181e+00
#>  [697,]    9.82066150 -2.231322e+00
#>  [698,]   12.25840107 -2.926170e+00
#>  [699,]   23.93612577 -5.569251e+00
#>  [700,]    7.53599720 -1.800190e+00
#>  [701,]    7.39538317 -1.498746e+00
#>  [702,]    7.12287794 -1.545000e+00
#>  [703,]    9.24712261 -2.061512e+00
#>  [704,]    2.85021024 -6.057965e-01
#>  [705,]   15.48272105 -3.712708e+00
#>  [706,]   11.76040397 -2.589498e+00
#>  [707,]   20.33015318 -4.890408e+00
#>  [708,]   26.48294221 -6.212104e+00
#>  [709,]   29.83527842 -7.006107e+00
#>  [710,]   14.28823109 -3.009371e+00
#>  [711,]    2.76323402 -5.519047e-01
#>  [712,]    6.99983598 -1.502032e+00
#>  [713,]   13.12466824 -3.022759e+00
#>  [714,]   27.76327470 -6.406029e+00
#>  [715,]    9.41357352 -2.139119e+00
#>  [716,]   10.38059589 -2.339093e+00
#>  [717,]   15.43013511 -3.701315e+00
#>  [718,]   10.22696757 -2.190360e+00
#>  [719,]    8.67597947 -1.942546e+00
#>  [720,]   15.18622301 -3.498010e+00
#>  [721,]    8.23640073 -1.989017e+00
#>  [722,]   14.75379053 -3.361380e+00
#>  [723,]   21.79856502 -4.885449e+00
#>  [724,]   11.79742978 -2.528290e+00
#>  [725,]   10.21880296 -2.473446e+00
#>  [726,]   11.17575130 -2.585023e+00
#>  [727,]    8.49736935 -2.035634e+00
#>  [728,]    2.38579510 -1.940925e-01
#>  [729,]   12.44836556 -2.856853e+00
#>  [730,]   17.05878030 -3.685722e+00
#>  [731,]   11.10579735 -2.626918e+00
#>  [732,]    7.64398151 -1.748325e+00
#>  [733,]   12.22443370 -2.722441e+00
#>  [734,]    6.56013520 -1.375716e+00
#>  [735,]   13.28430798 -3.022830e+00
#>  [736,]   17.16957376 -4.022300e+00
#>  [737,]    9.28565778 -2.267586e+00
#>  [738,]    7.74262498 -1.742406e+00
#>  [739,]    7.76822904 -1.678297e+00
#>  [740,]   11.14189831 -2.419624e+00
#>  [741,]    9.33252514 -2.082153e+00
#>  [742,]   12.75792798 -2.998892e+00
#>  [743,]   13.35679374 -3.232975e+00
#>  [744,]    7.52222742 -1.680827e+00
#>  [745,]    8.18833845 -1.669679e+00
#>  [746,]    7.88724867 -1.604855e+00
#>  [747,]   17.26336893 -4.210940e+00
#>  [748,]    8.59667844 -1.775175e+00
#>  [749,]   13.79566762 -3.123508e+00
#>  [750,]   15.00022451 -3.294087e+00
#>  [751,]   15.35881336 -3.449808e+00
#>  [752,]    7.68496483 -1.617800e+00
#>  [753,]   14.87955356 -3.271266e+00
#>  [754,]    6.23761024 -1.332660e+00
#>  [755,]   11.37159498 -2.318080e+00
#>  [756,]    2.92827507 -5.191942e-01
#>  [757,]    6.38576510 -1.273045e+00
#>  [758,]   19.40837043 -4.539861e+00
#>  [759,]   13.18195438 -3.004528e+00
#>  [760,]   15.53691894 -3.578412e+00
#>  [761,]    8.78497448 -1.878578e+00
#>  [762,]    8.50336355 -1.781263e+00
#>  [763,]   11.21474329 -2.577128e+00
#>  [764,]   14.28853577 -3.183470e+00
#>  [765,]    6.19485324 -1.342294e+00
#>  [766,]    9.65722598 -1.893546e+00
#>  [767,]    4.10093195 -7.373953e-01
#>  [768,]   12.22219503 -2.920311e+00
#>  [769,]   10.77132416 -2.330244e+00
#>  [770,]   10.74235141 -2.428608e+00
#>  [771,]    9.25685459 -2.056552e+00
#>  [772,]   12.75011489 -2.893050e+00
#>  [773,]    5.78136620 -1.099180e+00
#>  [774,]   16.32000983 -3.642009e+00
#>  [775,]    6.88147278 -1.527730e+00
#>  [776,]   11.93751628 -2.695952e+00
#>  [777,]   13.73589955 -3.126212e+00
#>  [778,]   12.09040760 -2.759506e+00
#>  [779,]   11.83798616 -2.887581e+00
#>  [780,]    6.87489684 -1.460885e+00
#>  [781,]    9.61484917 -2.273631e+00
#>  [782,]    5.93719402 -1.145537e+00
#>  [783,]    5.43294011 -1.012246e+00
#>  [784,]    6.04909099 -1.191604e+00
#>  [785,]   11.79037425 -2.652745e+00
#>  [786,]    9.27068148 -2.043152e+00
#>  [787,]   13.01770750 -3.050406e+00
#>  [788,]   21.08504249 -5.077377e+00
#>  [789,]   18.78800490 -4.414408e+00
#>  [790,]    7.12961327 -1.442795e+00
#>  [791,]    5.40636097 -1.164979e+00
#>  [792,]   14.53796952 -3.438875e+00
#>  [793,]   10.93343912 -2.657112e+00
#>  [794,]   13.08595700 -2.998545e+00
#>  [795,]   14.44013464 -3.122588e+00
#>  [796,]   16.04468237 -3.773239e+00
#>  [797,]    9.45799739 -2.087041e+00
#>  [798,]   17.80140241 -4.107627e+00
#>  [799,]    8.53476124 -1.972221e+00
#>  [800,]    6.93501173 -1.568822e+00
#>  [801,]    9.91582275 -2.102344e+00
#>  [802,]    4.48620902 -9.537518e-01
#>  [803,]   30.20713568 -7.186018e+00
#>  [804,]    1.18848431  5.045917e-02
#>  [805,]    5.89530509 -1.407943e+00
#>  [806,]   25.60780522 -6.226397e+00
#>  [807,]    8.09549208 -1.759576e+00
#>  [808,]    5.81263529 -1.124532e+00
#>  [809,]    6.94884101 -1.414646e+00
#>  [810,]   13.96987323 -2.989677e+00
#>  [811,]    2.39799950 -2.477226e-01
#>  [812,]    4.06409416 -6.626611e-01
#>  [813,]    6.57765540 -1.447064e+00
#>  [814,]   10.82399989 -2.452555e+00
#>  [815,]    9.69482588 -2.045014e+00
#>  [816,]    3.39378683 -7.156025e-01
#>  [817,]   17.70139047 -3.991741e+00
#>  [818,]   22.91237232 -5.534793e+00
#>  [819,]   16.62994673 -4.047562e+00
#>  [820,]   12.78039664 -2.915090e+00
#>  [821,]   11.42520634 -2.617979e+00
#>  [822,]   14.52020135 -3.295124e+00
#>  [823,]   13.82749861 -3.189179e+00
#>  [824,]    8.92163230 -2.171952e+00
#>  [825,]   14.96539546 -3.144689e+00
#>  [826,]   14.92972942 -3.564523e+00
#>  [827,]   18.62678586 -4.363595e+00
#>  [828,]   13.23033427 -3.104761e+00
#>  [829,]   10.08571185 -2.118780e+00
#>  [830,]   11.76446775 -2.567353e+00
#>  [831,]   20.49075955 -4.759978e+00
#>  [832,]    8.02442862 -1.732048e+00
#>  [833,]   12.39574540 -2.821769e+00
#>  [834,]   17.76730146 -4.307244e+00
#>  [835,]   15.26817175 -3.578721e+00
#>  [836,]    6.86936472 -1.418320e+00
#>  [837,]    9.21015674 -1.930008e+00
#>  [838,]   19.02275728 -4.464318e+00
#>  [839,]    7.59296574 -1.578366e+00
#>  [840,]   14.29771193 -3.261696e+00
#>  [841,]   27.41294045 -6.423926e+00
#>  [842,]    8.20929943 -1.860305e+00
#>  [843,]   -0.13262957  3.864719e-01
#>  [844,]   12.89850067 -2.894843e+00
#>  [845,]   11.42300929 -2.511869e+00
#>  [846,]    9.48764750 -2.039692e+00
#>  [847,]    5.00656588 -1.040899e+00
#>  [848,]   19.21757200 -4.574527e+00
#>  [849,]   86.30560341 -1.986164e+01
#>  [850,]   14.21838109 -3.187249e+00
#>  [851,]   16.28968923 -3.572304e+00
#>  [852,]   11.15486936 -2.399553e+00
#>  [853,]   12.05369967 -2.796996e+00
#>  [854,]   16.27572058 -3.749621e+00
#>  [855,]    0.14698062  3.161252e-01
#>  [856,]   10.01618868 -2.208851e+00
#>  [857,]    8.32544838 -1.694415e+00
#>  [858,]   12.41120568 -2.780399e+00
#>  [859,]    8.04909735 -1.742445e+00
#>  [860,]    9.70053228 -2.130596e+00
#>  [861,]   17.25097839 -3.883239e+00
#>  [862,]   12.42769848 -2.841529e+00
#>  [863,]    7.69041691 -1.706801e+00
#>  [864,]    4.65379365 -9.386815e-01
#>  [865,]   10.90322113 -2.442391e+00
#>  [866,]    7.35790418 -1.513016e+00
#>  [867,]   16.06713660 -4.063741e+00
#>  [868,]    8.93983568 -1.940193e+00
#>  [869,]   17.35909764 -3.903592e+00
#>  [870,]   22.08250417 -5.207093e+00
#>  [871,]    7.88369718 -1.572676e+00
#>  [872,]    1.37418672 -1.675294e-01
#>  [873,]    6.27727707 -1.430579e+00
#>  [874,]   18.53381236 -4.387335e+00
#>  [875,]   17.97382849 -4.102205e+00
#>  [876,]    9.91560461 -2.326460e+00
#>  [877,]    4.42361380 -8.969706e-01
#>  [878,]    7.78557757 -1.679176e+00
#>  [879,]    7.00358030 -1.336516e+00
#>  [880,]    4.85579393 -9.872399e-01
#>  [881,]   30.13249895 -7.031344e+00
#>  [882,]    6.24716792 -1.303553e+00
#>  [883,]   17.82991904 -4.016615e+00
#>  [884,]   14.89773712 -3.330287e+00
#>  [885,]   15.99246240 -3.807477e+00
#>  [886,]   10.21026780 -2.029643e+00
#>  [887,]    9.21764925 -2.134748e+00
#>  [888,]    9.56281431 -2.082691e+00
#>  [889,]   11.38080853 -2.560071e+00
#>  [890,]   14.45348894 -3.222331e+00
#>  [891,]   12.10645508 -2.938261e+00
#>  [892,]   10.35210716 -2.558460e+00
#>  [893,]    6.49611673 -1.291142e+00
#>  [894,]    9.50522110 -2.195010e+00
#>  [895,]    7.79843826 -1.590100e+00
#>  [896,]   18.08715713 -4.324945e+00
#>  [897,]   13.10140221 -3.048867e+00
#>  [898,]   21.86534589 -5.273920e+00
#>  [899,]    6.92191194 -1.452172e+00
#>  [900,]   12.44888729 -2.802682e+00
#>  [901,]   16.65480227 -4.146117e+00
#>  [902,]    7.09841490 -1.577634e+00
#>  [903,]   13.05929266 -3.046589e+00
#>  [904,]    9.93266162 -2.205264e+00
#>  [905,]    9.14554953 -2.003664e+00
#>  [906,]    9.69077016 -2.136575e+00
#>  [907,]   11.88182261 -2.637236e+00
#>  [908,]    8.98553665 -2.000777e+00
#>  [909,]   12.30522605 -2.763212e+00
#>  [910,]   14.32215869 -3.523145e+00
#>  [911,]   12.85133537 -2.862613e+00
#>  [912,]    5.04127914 -1.011342e+00
#>  [913,]    4.65186813 -1.120812e+00
#>  [914,]    6.71093032 -1.651091e+00
#>  [915,]   -2.55040384  9.821853e-01
#>  [916,]   11.46487429 -2.691651e+00
#>  [917,]    4.39394122 -6.994279e-01
#>  [918,]    6.48830736 -1.636900e+00
#>  [919,]   16.02355202 -3.584460e+00
#>  [920,]   25.27274602 -5.774859e+00
#>  [921,]   20.78150706 -4.804021e+00
#>  [922,]   14.54738047 -3.289651e+00
#>  [923,]   15.28981201 -3.342462e+00
#>  [924,]   15.55202718 -3.714937e+00
#>  [925,]   18.42056506 -4.182418e+00
#>  [926,]   10.43299369 -2.285334e+00
#>  [927,]   25.25534275 -5.272256e+00
#>  [928,]    2.38251117 -2.999075e-01
#>  [929,]   11.14327036 -2.476137e+00
#>  [930,]   32.35254644 -7.493813e+00
#>  [931,]   35.30484164 -7.966624e+00
#>  [932,]    9.09877845 -2.152042e+00
#>  [933,]    6.59074895 -1.188452e+00
#>  [934,]   10.36427357 -2.326286e+00
#>  [935,]   18.64085811 -4.481533e+00
#>  [936,]   18.45493090 -4.263369e+00
#>  [937,]   15.90330106 -3.650462e+00
#>  [938,]   14.87800029 -3.495350e+00
#>  [939,]    1.77443329 -2.342375e-01
#>  [940,]   15.74348166 -3.422403e+00
#>  [941,]    9.90739472 -2.161468e+00
#>  [942,]   14.86016428 -3.463397e+00
#>  [943,]    5.99169699 -1.323853e+00
#>  [944,]    3.38450678 -5.585345e-01
#>  [945,]   13.18848212 -3.037652e+00
#>  [946,]    6.28551394 -1.326084e+00
#>  [947,]    9.38566064 -2.028333e+00
#>  [948,]   11.58216031 -2.648573e+00
#>  [949,]   10.24068101 -2.181822e+00
#>  [950,]   30.32871963 -6.841759e+00
#>  [951,]   10.39358869 -2.300343e+00
#>  [952,]   21.62798669 -5.098611e+00
#>  [953,]   11.37411058 -2.649312e+00
#>  [954,]    8.54423251 -1.873125e+00
#>  [955,]    8.95171442 -2.014473e+00
#>  [956,]   10.08873771 -2.348467e+00
#>  [957,]   19.31352692 -4.577978e+00
#>  [958,]    4.34821119 -7.593380e-01
#>  [959,]   14.34191232 -3.547577e+00
#>  [960,]    1.71652647 -4.290606e-01
#>  [961,]   26.09855046 -6.304729e+00
#>  [962,]    4.15479697 -6.935236e-01
#>  [963,]   16.14101877 -3.634437e+00
#>  [964,]    6.76494096 -1.366689e+00
#>  [965,]   12.10893509 -2.712604e+00
#>  [966,]    3.56339454 -7.210906e-01
#>  [967,]   17.92132980 -4.147133e+00
#>  [968,]   20.55134558 -4.971299e+00
#>  [969,]   14.26235745 -3.221190e+00
#>  [970,]   13.31986547 -3.003924e+00
#>  [971,]   17.85575082 -4.165143e+00
#>  [972,]    6.32523528 -1.541344e+00
#>  [973,]    5.25413034 -1.179544e+00
#>  [974,]    6.90892802 -1.567247e+00
#>  [975,]   18.02251888 -4.203377e+00
#>  [976,]   28.23172812 -6.672420e+00
#>  [977,]   15.14152293 -3.462399e+00
#>  [978,]   17.16678456 -3.984457e+00
#>  [979,]    0.61762524 -1.799526e-02
#>  [980,]   15.86068498 -3.782810e+00
#>  [981,]   13.42315982 -3.065283e+00
#>  [982,]    7.38334070 -1.786202e+00
#>  [983,]    7.35525573 -1.496455e+00
#>  [984,]   10.23120557 -2.227009e+00
#>  [985,]   13.45885645 -3.063417e+00
#>  [986,]    5.82296578 -1.141986e+00
#>  [987,]   25.19908401 -5.870173e+00
#>  [988,]   12.26482694 -2.982153e+00
#>  [989,]   10.68903443 -2.278720e+00
#>  [990,]   12.36271176 -2.887792e+00
#>  [991,]   10.52247554 -2.570335e+00
#>  [992,]    5.26739976 -9.257696e-01
#>  [993,]   11.67008954 -2.609984e+00
#>  [994,]    6.89595123 -1.445204e+00
#>  [995,]   14.62511546 -3.543891e+00
#>  [996,]   15.76916280 -3.569251e+00
#>  [997,]   15.83884907 -3.545305e+00
#>  [998,]   11.82891721 -2.643493e+00
#>  [999,]    6.84826946 -1.539747e+00
#> [1000,]    2.53773939 -3.381357e-01
#> [1001,]   20.19509697 -4.998037e+00
#> [1002,]   12.21037751 -2.932893e+00
#> [1003,]   13.26068210 -2.962142e+00
#> [1004,]    3.95220188 -7.239739e-01
#> [1005,]    9.64948728 -1.973173e+00
#> [1006,]   12.24765032 -2.868509e+00
#> [1007,]   23.52887745 -5.704358e+00
#> [1008,]    9.16833345 -2.047994e+00
#> [1009,]    2.54618446 -4.160888e-01
#> [1010,]   17.45824055 -4.000977e+00
#> [1011,]   12.60275054 -2.823595e+00
#> [1012,]   22.02804339 -5.039411e+00
#> [1013,]   14.77647387 -3.197425e+00
#> [1014,]    8.77069821 -2.081934e+00
#> [1015,]   12.71497417 -2.805026e+00
#> [1016,]   14.41787525 -3.310056e+00
#> [1017,]   22.04008145 -5.363092e+00
#> [1018,]   10.40571022 -2.392759e+00
#> [1019,]   14.78591370 -3.417962e+00
#> [1020,]   16.45891569 -3.974214e+00
#> [1021,]   12.79131564 -2.951573e+00
#> [1022,]   18.12688016 -4.289027e+00
#> [1023,]   24.36716323 -5.645068e+00
#> [1024,]   10.08948263 -2.352873e+00
#> [1025,]    2.08262814 -2.614982e-01
#> [1026,]   16.44354503 -3.689066e+00
#> [1027,]   10.75321689 -2.361327e+00
#> [1028,]    6.72434312 -1.163434e+00
#> [1029,]  100.25394508 -2.336538e+01
#> [1030,]    6.77003424 -1.430036e+00
#> [1031,]   15.63201963 -3.809202e+00
#> [1032,]    7.63339688 -1.797642e+00
#> [1033,]   10.24791724 -2.365509e+00
#> [1034,]   12.77998524 -2.944899e+00
#> [1035,]    5.95822354 -1.297839e+00
#> [1036,]   13.72080669 -3.092300e+00
#> [1037,]   10.24313189 -2.294695e+00
#> [1038,]   10.01686767 -2.357276e+00
#> [1039,]   13.32311070 -3.046295e+00
#> [1040,]    9.39900778 -1.964319e+00
#> [1041,]   11.06373340 -2.610723e+00
#> [1042,]   17.04335543 -3.861264e+00
#> [1043,]   12.48312253 -2.858492e+00
#> [1044,]    9.70614957 -1.977651e+00
#> [1045,]    9.28037053 -2.255773e+00
#> [1046,]   13.63356975 -2.916319e+00
#> [1047,]   10.23932047 -2.369631e+00
#> [1048,]   11.62337188 -2.715048e+00
#> [1049,]   16.93246823 -4.007055e+00
#> [1050,]   12.25635720 -2.797972e+00
#> [1051,]   13.30045293 -3.047796e+00
#> [1052,]    7.47847047 -1.655333e+00
#> [1053,]    5.15396231 -1.042574e+00
#> [1054,]    8.41894723 -1.859752e+00
#> [1055,]    7.14414965 -1.603700e+00
#> [1056,]   18.42930721 -4.441695e+00
#> [1057,]   13.05673089 -2.806425e+00
#> [1058,]   12.05573890 -2.759179e+00
#> [1059,]   17.39810142 -3.972629e+00
#> [1060,]   21.69497919 -5.198202e+00
#> [1061,]    9.90043067 -2.109800e+00
#> [1062,]   24.33702674 -5.830375e+00
#> [1063,]   13.12665978 -3.079441e+00
#> [1064,]    4.38821516 -7.905081e-01
#> [1065,]    7.87049241 -1.792375e+00
#> [1066,]   12.05683762 -2.749637e+00
#> [1067,]   13.26014552 -2.926027e+00
#> [1068,]   17.85283650 -4.094930e+00
#> [1069,]   22.95270673 -5.350880e+00
#> [1070,]    5.89860892 -1.424112e+00
#> [1071,]   20.29245246 -5.149496e+00
#> [1072,]    7.63001860 -1.553106e+00
#> [1073,]   14.27182432 -3.148156e+00
#> [1074,]    5.84982134 -1.113748e+00
#> [1075,]    9.00616838 -1.774952e+00
#> [1076,]    7.35261455 -1.642964e+00
#> [1077,]    8.00843293 -1.782819e+00
#> [1078,]   13.72059255 -3.190522e+00
#> [1079,]    2.99839310 -4.933054e-01
#> [1080,]   10.54353234 -2.318356e+00
#> [1081,]    6.53833611 -1.416738e+00
#> [1082,]    8.25148834 -1.906211e+00
#> [1083,]   20.02549271 -4.758216e+00
#> [1084,]   18.42471366 -4.445143e+00
#> [1085,]   11.11453401 -2.581882e+00
#> [1086,]    8.45407367 -1.664674e+00
#> [1087,]   13.35149052 -3.053293e+00
#> [1088,]   14.77468091 -3.303458e+00
#> [1089,]    8.21515314 -1.711446e+00
#> [1090,]    5.98200706 -1.295596e+00
#> [1091,]   11.01614402 -2.426765e+00
#> [1092,]   10.78386373 -2.276318e+00
#> [1093,]   14.91360419 -3.640181e+00
#> [1094,]   14.29916961 -3.505301e+00
#> [1095,]    7.16913432 -1.559826e+00
#> [1096,]    7.67494854 -1.771776e+00
#> [1097,]    7.56481972 -1.561008e+00
#> [1098,]    6.53833011 -1.422673e+00
#> [1099,]    5.57031736 -1.187098e+00
#> [1100,]    6.49075384 -1.205599e+00
#> [1101,]   11.76532787 -2.645221e+00
#> [1102,]   13.53704504 -3.213712e+00
#> [1103,]    7.66169969 -1.727667e+00
#> [1104,]    8.97630931 -1.988788e+00
#> [1105,]    7.31881164 -1.555411e+00
#> [1106,]   12.17104303 -2.768360e+00
#> [1107,]   16.99516085 -4.001335e+00
#> [1108,]   12.79201675 -3.024605e+00
#> [1109,]   11.41146516 -2.441536e+00
#> [1110,]   11.00826353 -2.555123e+00
#> [1111,]   11.48740460 -2.629003e+00
#> [1112,]    9.48386052 -2.191648e+00
#> [1113,]   14.52293438 -3.677512e+00
#> [1114,]    5.13044088 -1.041590e+00
#> [1115,]    8.51862919 -1.828730e+00
#> [1116,]    2.04858535 -1.661225e-01
#> [1117,]    8.88210990 -2.078368e+00
#> [1118,]    8.83181684 -1.994391e+00
#> [1119,]    9.75848224 -2.303870e+00
#> [1120,]    6.77414245 -1.541607e+00
#> [1121,]    9.81082783 -2.085426e+00
#> [1122,]    5.97308601 -1.308841e+00
#> [1123,]    9.53962869 -2.333574e+00
#> [1124,]   13.10200331 -3.178069e+00
#> [1125,]   17.74447832 -4.326999e+00
#> [1126,]   19.37990912 -4.648992e+00
#> [1127,]   14.79118474 -3.444041e+00
#> [1128,]   11.49851346 -2.768111e+00
#> [1129,]    7.46392797 -1.582432e+00
#> [1130,]   10.63381260 -2.336748e+00
#> [1131,]    3.82312973 -6.565111e-01
#> [1132,]   13.61267705 -3.290341e+00
#> [1133,]   16.38764478 -3.805829e+00
#> [1134,]   10.25382499 -2.553919e+00
#> [1135,]    9.41188544 -2.256959e+00
#> [1136,]    9.16239868 -2.004572e+00
#> [1137,]    5.03899430 -8.386363e-01
#> [1138,]   15.26424010 -3.284322e+00
#> [1139,]   12.39234272 -2.723777e+00
#> [1140,]   13.53995453 -3.015138e+00
#> [1141,]    5.00220097 -1.088182e+00
#> [1142,]    8.85845541 -2.074879e+00
#> [1143,]   19.66330544 -4.340703e+00
#> [1144,]    7.26838177 -1.521624e+00
#> [1145,]    4.35209865 -9.081852e-01
#> [1146,]   12.62951042 -3.014514e+00
#> [1147,]   12.37920042 -2.903804e+00
#> [1148,]   11.12049238 -2.588133e+00
#> [1149,]   10.83607653 -2.354343e+00
#> [1150,]   13.56985319 -3.217256e+00
#> [1151,]    6.21071259 -1.264510e+00
#> [1152,]    1.28645986 -1.866759e-01
#> [1153,]   10.49253239 -2.423282e+00
#> [1154,]    8.56893550 -1.738892e+00
#> [1155,]   12.77179448 -3.011883e+00
#> [1156,]    8.59762537 -1.909003e+00
#> [1157,]   23.69873883 -5.773069e+00
#> [1158,]   23.58676574 -5.195800e+00
#> [1159,]   10.81671537 -2.420844e+00
#> [1160,]    7.23494642 -1.589856e+00
#> [1161,]   14.89500133 -3.523837e+00
#> [1162,]   13.05364111 -3.064891e+00
#> [1163,]   14.06779301 -3.138502e+00
#> [1164,]   16.94334757 -3.998608e+00
#> [1165,]    9.84627837 -2.170368e+00
#> [1166,]    9.42734833 -2.109586e+00
#> [1167,]   39.21256848 -9.308031e+00
#> [1168,]   15.65658880 -3.737121e+00
#> [1169,]    9.34151570 -1.992027e+00
#> [1170,]    7.00621558 -1.510183e+00
#> [1171,]   18.48533781 -4.240244e+00
#> [1172,]   15.52960069 -3.648451e+00
#> [1173,]   11.13842191 -2.672160e+00
#> [1174,]   12.92384955 -2.954001e+00
#> [1175,]    2.43501302 -3.150928e-01
#> [1176,]   14.66236515 -3.100692e+00
#> [1177,]   14.19745453 -3.385067e+00
#> [1178,]   11.17565065 -2.676753e+00
#> [1179,]   21.62704358 -5.232670e+00
#> [1180,]   12.96491508 -2.869972e+00
#> [1181,]   15.29499143 -3.576696e+00
#> [1182,]    7.46514922 -1.701865e+00
#> [1183,]    5.55284329 -1.244522e+00
#> [1184,]   10.92658534 -2.491194e+00
#> [1185,]    6.30279452 -1.202215e+00
#> [1186,]   12.16593800 -2.767302e+00
#> [1187,]   31.36182558 -7.394483e+00
#> [1188,]   12.11782063 -2.716607e+00
#> [1189,]   10.56916543 -2.360493e+00
#> [1190,]   15.97662644 -3.647541e+00
#> [1191,]   21.28497723 -4.954553e+00
#> [1192,]    7.98353379 -1.831897e+00
#> [1193,]    7.48055545 -1.587765e+00
#> [1194,]   12.11043001 -2.833894e+00
#> [1195,]    2.65492456 -4.125334e-01
#> [1196,]   14.59142969 -3.354802e+00
#> [1197,]    4.85045448 -9.749432e-01
#> [1198,]   11.28826413 -2.599484e+00
#> [1199,]    7.58696362 -1.660505e+00
#> [1200,]    8.70177018 -2.038903e+00
#> [1201,]    4.79635290 -8.247613e-01
#> [1202,]   11.82690140 -2.648730e+00
#> [1203,]    8.37229353 -1.911477e+00
#> [1204,]   14.74704363 -3.450976e+00
#> [1205,]    8.02275876 -1.754439e+00
#> [1206,]   11.52547665 -2.486816e+00
#> [1207,]   10.09200062 -2.242676e+00
#> [1208,]    5.07122279 -1.009831e+00
#> [1209,]   15.42372424 -3.327306e+00
#> [1210,]   26.63600838 -6.448408e+00
#> [1211,]    7.56681035 -1.664144e+00
#> [1212,]    7.42485546 -1.585781e+00
#> [1213,]    8.36253175 -1.913727e+00
#> [1214,]   -0.75170067  2.224582e-01
#> [1215,]    8.36355484 -1.894700e+00
#> [1216,]    6.03285404 -1.258212e+00
#> [1217,]    7.42927526 -1.608170e+00
#> [1218,]    3.95077395 -7.345539e-01
#> [1219,]   10.39313638 -2.325926e+00
#> [1220,]    8.97730472 -1.925353e+00
#> [1221,]   16.54288612 -3.948668e+00
#> [1222,]    8.78540138 -1.783939e+00
#> [1223,]    9.48084745 -2.209144e+00
#> [1224,]    9.59036917 -2.228328e+00
#> [1225,]   13.76275551 -2.925784e+00
#> [1226,]    7.87104117 -1.669466e+00
#> [1227,]   15.05040065 -3.454389e+00
#> [1228,]    1.88615362 -3.709141e-01
#> [1229,]   12.11750729 -2.772215e+00
#> [1230,]   19.55825610 -4.757626e+00
#> [1231,]    7.08295436 -1.379943e+00
#> [1232,]    0.07392656  3.881565e-01
#> [1233,]   11.27095535 -2.616868e+00
#> [1234,]   15.86810543 -3.629332e+00
#> [1235,]   15.25202418 -3.635522e+00
#> [1236,]    9.04349038 -2.044207e+00
#> [1237,]    9.12568609 -2.087606e+00
#> [1238,]    8.18984330 -1.677027e+00
#> [1239,]    9.10769137 -2.082604e+00
#> [1240,]   14.12173192 -3.316662e+00
#> [1241,]   13.72995190 -3.016227e+00
#> [1242,]   13.22865336 -3.010132e+00
#> [1243,]    9.26273386 -2.147341e+00
#> [1244,]   14.36857788 -3.189045e+00
#> [1245,]   29.19090438 -6.502562e+00
#> [1246,]   26.03011194 -6.066438e+00
#> [1247,]   12.84292672 -2.881299e+00
#> [1248,]    9.25655676 -1.790566e+00
#> [1249,]    7.69411014 -1.686899e+00
#> [1250,]   11.09116472 -2.477125e+00
#> [1251,]    4.30588420 -8.530922e-01
#> [1252,]    4.40666892 -7.684691e-01
#> [1253,]    4.67787934 -9.050412e-01
#> [1254,]   12.94440646 -2.834169e+00
#> [1255,]   10.89202538 -2.633090e+00
#> [1256,]   15.16967701 -3.570483e+00
#> [1257,]   15.49798780 -3.242848e+00
#> [1258,]   10.74330425 -2.375723e+00
#> [1259,]    4.90011442 -9.357694e-01
#> [1260,]   13.76050203 -3.126075e+00
#> [1261,]    4.19785237 -7.032881e-01
#> [1262,]    6.57094696 -1.464834e+00
#> [1263,]   14.17355289 -3.213338e+00
#> [1264,]   10.00151614 -2.330864e+00
#> [1265,]    7.28744149 -1.529501e+00
#> [1266,]   14.75004008 -3.106279e+00
#> [1267,]    9.18967539 -2.153468e+00
#> [1268,]   11.92143613 -2.555590e+00
#> [1269,]    8.59278290 -1.819591e+00
#> [1270,]   19.53750638 -4.710093e+00
#> [1271,]    9.92846261 -2.055297e+00
#> [1272,]   13.40126840 -3.068020e+00
#> [1273,]    2.01989200 -1.595069e-01
#> [1274,]    7.31000987 -1.685494e+00
#> [1275,]   17.64181793 -4.222851e+00
#> [1276,]   15.17914547 -3.448396e+00
#> [1277,]    6.20798072 -1.304339e+00
#> [1278,]    5.99642503 -1.398211e+00
#> [1279,]   10.09033793 -2.208408e+00
#> [1280,]   10.29654592 -2.310834e+00
#> [1281,]    6.85925529 -1.319178e+00
#> [1282,]   10.75263558 -2.118095e+00
#> [1283,]   10.76930297 -2.521053e+00
#> [1284,]   13.93120415 -3.267041e+00
#> [1285,]    8.07285435 -1.732055e+00
#> [1286,]   16.33342603 -3.889756e+00
#> [1287,]   -1.76161927  6.014397e-01
#> [1288,]   11.56542303 -2.425418e+00
#> [1289,]   20.46833983 -4.983367e+00
#> [1290,]   25.16025954 -5.717492e+00
#> [1291,]   17.84704643 -4.229871e+00
#> [1292,]   13.47062738 -2.962634e+00
#> [1293,]   17.41723033 -4.080564e+00
#> [1294,]   11.20268016 -2.455490e+00
#> [1295,]   20.93394327 -4.787160e+00
#> [1296,]    9.06865249 -2.040430e+00
#> [1297,]    7.23443531 -1.434727e+00
#> [1298,]   23.84428826 -5.519009e+00
#> [1299,]    7.63327901 -1.731766e+00
#> [1300,]    5.87401102 -1.296856e+00
#> [1301,]    6.76429358 -1.473763e+00
#> [1302,]   15.55505866 -3.594421e+00
#> [1303,]   13.39422575 -2.826704e+00
#> [1304,]   16.32118860 -3.794027e+00
#> [1305,]   15.21837529 -3.629697e+00
#> [1306,]    8.60281697 -1.979486e+00
#> [1307,]    5.56476273 -9.328977e-01
#> [1308,]    2.42237996 -3.016007e-01
#> [1309,]   23.11724791 -5.392003e+00
#> [1310,]   12.24674414 -2.796726e+00
#> [1311,]   11.26338044 -2.420759e+00
#> [1312,]   10.96805447 -2.651438e+00
#> [1313,]    8.26155028 -1.835251e+00
#> [1314,]   24.75023219 -5.534214e+00
#> [1315,]   20.85078686 -4.864871e+00
#> [1316,]   10.95370433 -2.429311e+00
#> [1317,]   41.50926582 -9.842281e+00
#> [1318,]   11.37517854 -2.321066e+00
#> [1319,]   26.27927814 -6.021410e+00
#> [1320,]    9.15789156 -2.140482e+00
#> [1321,]   17.15035785 -4.123634e+00
#> [1322,]    5.08413278 -1.085257e+00
#> [1323,]    4.62443165 -1.023993e+00
#> [1324,]   13.16798759 -3.033099e+00
#> [1325,]   16.76875065 -3.720921e+00
#> [1326,]   11.08170009 -2.501573e+00
#> [1327,]    9.53709401 -2.333364e+00
#> [1328,]   14.22019293 -3.127396e+00
#> [1329,]   10.28491995 -2.313722e+00
#> [1330,]   12.47745579 -2.983183e+00
#> [1331,]   15.53506450 -3.597502e+00
#> [1332,]    9.75745664 -1.946176e+00
#> [1333,]    8.70613173 -1.882709e+00
#> [1334,]    9.72879649 -2.207599e+00
#> [1335,]   15.85883521 -3.717616e+00
#> [1336,]    8.42893010 -2.055419e+00
#> [1337,]    8.66743751 -1.854245e+00
#> [1338,]    9.96839583 -2.400798e+00
#> [1339,]    8.40702919 -1.817243e+00
#> [1340,]   11.09289356 -2.555801e+00
#> [1341,]   10.18176185 -2.330873e+00
#> [1342,]    7.14503923 -1.638628e+00
#> [1343,]    6.27871360 -1.229742e+00
#> [1344,]    4.55721675 -7.491761e-01
#> [1345,]    4.22291124 -8.038957e-01
#> [1346,]    3.38895068 -5.085318e-01
#> [1347,]    5.63034295 -1.122832e+00
#> [1348,]   21.51936940 -5.161876e+00
#> [1349,]   10.36160725 -2.386314e+00
#> [1350,]   10.10326201 -2.318876e+00
#> [1351,]   18.16241442 -4.145667e+00
#> [1352,]    5.91724716 -1.126979e+00
#> [1353,]   13.80862012 -3.095886e+00
#> [1354,]    3.11970235 -6.348244e-01
#> [1355,]    9.08234433 -2.056567e+00
#> [1356,]   16.23039586 -3.969522e+00
#> [1357,]   11.59271012 -2.611843e+00
#> [1358,]   15.55235456 -3.590289e+00
#> [1359,]   12.48178243 -2.856670e+00
#> [1360,]   12.80175708 -2.879408e+00
#> [1361,]   21.19542522 -5.271408e+00
#> [1362,]   11.53336199 -2.694736e+00
#> [1363,]    6.49808900 -1.267243e+00
#> [1364,]   21.45283856 -5.031211e+00
#> [1365,]    2.46523388 -2.748988e-01
#> [1366,]   16.93398457 -3.863582e+00
#> [1367,]   30.92048996 -7.167630e+00
#> [1368,]   31.01495324 -7.422515e+00
#> [1369,]   16.06087712 -3.876056e+00
#> [1370,]    7.56290880 -1.709453e+00
#> [1371,]   10.39987668 -2.444275e+00
#> [1372,]   11.90642126 -2.682767e+00
#> [1373,]    7.96164261 -1.831148e+00
#> [1374,]   13.47614224 -3.026036e+00
#> [1375,]    3.83479996 -7.578325e-01
#> [1376,]   18.11004435 -4.157828e+00
#> [1377,]   85.16559542 -1.960885e+01
#> [1378,]    8.14632315 -1.744925e+00
#> [1379,]   17.65880869 -4.018658e+00
#> [1380,]    6.57663436 -1.351280e+00
#> [1381,]    2.50625158 -2.832830e-01
#> [1382,]    3.71747745 -6.872670e-01
#> [1383,]   10.83043777 -2.274235e+00
#> [1384,]   20.59041318 -4.726911e+00
#> [1385,]   15.94017760 -3.548197e+00
#> [1386,]   14.20240209 -3.254854e+00
#> [1387,]   15.26508253 -3.425250e+00
#> [1388,]   11.81964142 -2.814284e+00
#> [1389,]    9.81221744 -2.162252e+00
#> [1390,]   22.97940883 -5.503531e+00
#> [1391,]   25.20243745 -6.034734e+00
#> [1392,]   11.33560559 -2.696502e+00
#> [1393,]   15.42360426 -3.671243e+00
#> [1394,]    1.35318560 -1.662078e-01
#> [1395,]   20.17161063 -4.508443e+00
#> [1396,]   15.72324351 -3.520425e+00
#> [1397,]   11.43207104 -2.775366e+00
#> [1398,]    5.07254849 -1.173732e+00
#> [1399,]   11.97937279 -2.708198e+00
#> [1400,]   10.94081817 -2.377318e+00
#> 
#> $varmethod
#> [1] "Bootstrap"
#> 
#> attr(,"class")
#> [1] "get_detection"